Split generic specialization generated C tests
