+++ docs/COMPLETION_AUDIT.md
@@ -2986,6 +2986,13 @@ and do not assume Phase 4 is ready without evidence.
   `src/resolver/symbol_table_test_support/aggregate_metadata.rs`, preserving
   existing test helper APIs while keeping the parent test-support module
   focused on shared lookup/indexing and generic value metadata setters.
+- Single-file generic specialization generated-C assertions now keep enum
+  specialization coverage in
+  `tests/integration/generic_specializations/enum_generated_c.rs` and
+  method/worklist specialization coverage in
+  `tests/integration/generic_specializations/method_worklist_generated_c.rs`,
+  preserving undefined-call and unspecialized-symbol checks while keeping the
+  parent integration module focused on cross-fixture uniqueness.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
