+++ tests/integration/generic_specializations.rs
@@ -2,6 +2,10 @@ use super::support::*;
 
 #[path = "generic_specializations/behavior_bounds.rs"]
 mod behavior_bounds;
+#[path = "generic_specializations/enum_generated_c.rs"]
+mod enum_generated_c;
+#[path = "generic_specializations/method_worklist_generated_c.rs"]
+mod method_worklist_generated_c;
 #[path = "generic_specializations/multifile_generated_c.rs"]
 mod multifile_generated_c;
 
@@ -43,222 +47,4 @@ fn generic_specializations_emit_each_generated_c_definition_once() {
         let c_source = compile_to_c_with_generated_call_check(&test_dir().join(fixture));
         assert_generated_c_function_definitions_are_unique(&c_source);
     }
-}
-
-#[test]
-fn generic_specializations_do_not_emit_unspecialized_c_symbols() {
-    let c_source = compile_to_c_with_generated_call_check(&test_dir().join("generic_method.zen"));
-    assert!(c_source.contains("int32_t Box_get_i32(Box_i32 self)"));
-    assert!(c_source.contains("Box_get_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_get_i32");
-    assert!(!c_source.contains("Box_T"));
-    assert!(!c_source.contains("T Box_get"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_method_self.zen"));
-    assert!(c_source.contains("Box_i32 Box_copy_i32(Box_i32 self)"));
-    assert!(c_source.contains("Box_copy_i32(box)"));
-    assert!(c_source.contains("Box_Option_i32 Box_copy_Option_i32(Box_Option_i32 self)"));
-    assert!(c_source.contains("Box_copy_Option_i32(nested)"));
-    assert!(c_source.contains("Option_i32 Option_copy_i32(Option_i32 self)"));
-    assert!(c_source.contains("Option_copy_i32(option)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_copy_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_copy_Option_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Option_copy_i32");
-    assert!(
-        c_source
-            .find("struct Option_i32")
-            .expect("Option_i32 struct")
-            < c_source
-                .find("struct Box_Option_i32")
-                .expect("Box_Option_i32 struct")
-    );
-    assert!(!c_source.contains("Box_copy(box"));
-    assert!(!c_source.contains("Unknown"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_method_worklist.zen"));
-    assert!(c_source.contains("int32_t inner_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t Box_get_inner_i32(Box_i32 self)"));
-    assert!(c_source.contains("inner_i32(self.value)"));
-    assert!(c_source.contains("Box_get_inner_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_get_inner_i32");
-    assert!(!c_source.contains("T inner"));
-    assert!(!c_source.contains("inner_T"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("generic_method_nested_result.zen"),
-    );
-    assert!(c_source.contains("typedef struct Result_Option_i32_str Result_Option_i32_str;"));
-    assert!(c_source.contains("Result_Option_i32_str Box_wrap_result_i32(Box_i32 self)"));
-    assert!(c_source.contains("Box_wrap_result_i32(box)"));
-    assert!(c_source.contains("unwrap_result_Option_i32_str(wrapped,"));
-    assert!(c_source.contains("unwrap_option_i32(some, 0LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_wrap_result_i32");
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_result_Option_i32_str");
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_option_i32");
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T Box_wrap_result"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("type_impl_methods.zen"));
-    assert!(c_source.contains("int32_t Point_get(Point self)"));
-    assert!(c_source.contains("int32_t Point_keep_i32(Point self, int32_t value)"));
-    assert!(c_source.contains("Point_get(point)"));
-    assert!(c_source.contains("Point_keep_i32(point, 7LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Point_get");
-    assert_c_call_resolves_to_definition(&c_source, "Point_keep_i32");
-    assert!(!c_source.contains("T Point_keep"));
-    assert!(!c_source.contains("Point_keep(point"));
-
-    let c_source = compile_to_c_with_generated_call_check(&test_dir().join("generic_vec.zen"));
-    assert!(c_source.contains("int32_t Vec_len_i32(Vec_i32 self)"));
-    assert!(c_source.contains("int32_t Vec_len_str(Vec_str self)"));
-    assert!(c_source.contains("Vec_len_i32(ints)"));
-    assert!(c_source.contains("Vec_len_str(words)"));
-    assert_c_call_resolves_to_definition(&c_source, "Vec_len_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Vec_len_str");
-    assert!(!c_source.contains("Vec_T"));
-    assert!(!c_source.contains("T Vec_len"));
-
-    let c_source = compile_to_c_with_generated_call_check(&test_dir().join("generic_worklist.zen"));
-    assert!(c_source.contains("int32_t inner_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t outer_i32(int32_t value)"));
-    assert!(c_source.contains("inner_i32(value)"));
-    assert_c_call_resolves_to_definition(&c_source, "inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "outer_i32");
-    assert_c_function_definition_count(&c_source, "inner_i32", 1);
-    assert!(!c_source.contains("T inner"));
-    assert!(!c_source.contains("inner_T"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_worklist_dedup.zen"));
-    assert!(c_source.contains("int32_t left_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t right_i32(int32_t value)"));
-    assert!(c_source.contains("inner_i32(value)"));
-    assert_c_call_resolves_to_definition(&c_source, "inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "left_i32");
-    assert_c_call_resolves_to_definition(&c_source, "right_i32");
-    assert_c_function_definition_count(&c_source, "inner_i32", 1);
-    assert!(!c_source.contains("T inner"));
-    assert!(!c_source.contains("inner_T"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_enum_option.zen"));
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("int32_t unwrap_or_i32(Option_i32 value, int32_t fallback)"));
-    assert!(c_source.contains("Option_i32_Some"));
-    assert!(c_source.contains("unwrap_or_i32(x, 0LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_or_i32");
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T unwrap_or"));
-    assert!(!c_source.contains("unwrap_or(x"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_enum_method.zen"));
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("int32_t Option_unwrap_or_i32(Option_i32 self, int32_t fallback)"));
-    assert!(c_source.contains("Option_unwrap_or_i32(some, 0LL)"));
-    assert!(c_source.contains("Option_unwrap_or_i32(none, 55LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Option_unwrap_or_i32");
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T Option_unwrap_or"));
-    assert!(!c_source.contains("Option_unwrap_or(some"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("generic_enum_multi_specialization.zen"),
-    );
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("typedef struct Option_bool Option_bool;"));
-    assert!(c_source.contains("int32_t Option_unwrap_or_i32(Option_i32 self, int32_t fallback)"));
-    assert!(c_source.contains("bool Option_unwrap_or_bool(Option_bool self, bool fallback)"));
-    assert!(c_source.contains("Option_unwrap_or_i32(some_int, 0LL)"));
-    assert!(c_source.contains("Option_unwrap_or_i32(none_int, 55LL)"));
-    assert!(c_source.contains("Option_unwrap_or_bool(some_bool, false)"));
-    assert!(c_source.contains("Option_unwrap_or_bool(none_bool, true)"));
-    assert_c_call_resolves_to_definition(&c_source, "Option_unwrap_or_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Option_unwrap_or_bool");
-    assert_c_function_definition_count(&c_source, "Option_unwrap_or_i32", 1);
-    assert_c_function_definition_count(&c_source, "Option_unwrap_or_bool", 1);
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T Option_unwrap_or"));
-    assert!(!c_source.contains("Option_unwrap_or(some"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("duplicate_enum_variant_names.zen"),
-    );
-    assert!(c_source.contains("First_i32_Some"));
-    assert!(c_source.contains("First_i32_None"));
-    assert!(c_source.contains("Second_bool_Some"));
-    assert!(c_source.contains("Second_bool_None"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_result_enum.zen"));
-    assert!(c_source.contains("typedef struct Result_i32_str Result_i32_str;"));
-    assert!(c_source.contains("int32_t unwrap_or_i32_str(Result_i32_str value, int32_t fallback)"));
-    assert!(c_source.contains("Result_i32_str_Err"));
-    assert!(c_source.contains("unwrap_or_i32_str(err, 9LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_or_i32_str");
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("T unwrap_or"));
-    assert!(!c_source.contains("unwrap_or(err"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_result_enum_method.zen"));
-    assert!(c_source.contains("typedef struct Result_i32_str Result_i32_str;"));
-    assert!(c_source
-        .contains("int32_t Result_unwrap_or_i32_str(Result_i32_str self, int32_t fallback)"));
-    assert!(c_source.contains("Result_unwrap_or_i32_str(ok, 0LL)"));
-    assert!(c_source.contains("Result_unwrap_or_i32_str(err, 34LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Result_unwrap_or_i32_str");
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("T Result_unwrap_or"));
-    assert!(!c_source.contains("Result_unwrap_or(err"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("generic_result_enum_multi_specialization.zen"),
-    );
-    assert!(c_source.contains("typedef struct Result_i32_str Result_i32_str;"));
-    assert!(c_source.contains("typedef struct Result_bool_str Result_bool_str;"));
-    assert!(c_source
-        .contains("int32_t Result_unwrap_or_i32_str(Result_i32_str self, int32_t fallback)"));
-    assert!(
-        c_source.contains("bool Result_unwrap_or_bool_str(Result_bool_str self, bool fallback)")
-    );
-    assert!(c_source.contains("Result_unwrap_or_i32_str(ok_int, 0LL)"));
-    assert!(c_source.contains("Result_unwrap_or_i32_str(err_int, 34LL)"));
-    assert!(c_source.contains("Result_unwrap_or_bool_str(ok_bool, true)"));
-    assert!(c_source.contains("Result_unwrap_or_bool_str(err_bool, true)"));
-    assert_c_call_resolves_to_definition(&c_source, "Result_unwrap_or_i32_str");
-    assert_c_call_resolves_to_definition(&c_source, "Result_unwrap_or_bool_str");
-    assert_c_function_definition_count(&c_source, "Result_unwrap_or_i32_str", 1);
-    assert_c_function_definition_count(&c_source, "Result_unwrap_or_bool_str", 1);
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("T Result_unwrap_or"));
-    assert!(!c_source.contains("Result_unwrap_or(err"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_nested_result_enum.zen"));
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("typedef struct Result_Option_i32_str Result_Option_i32_str;"));
-    assert!(c_source.contains(
-        "Option_i32 unwrap_result_Option_i32_str(Result_Option_i32_str value, Option_i32 fallback)"
-    ));
-    assert!(c_source.contains("unwrap_result_Option_i32_str(ok,"));
-    assert!(c_source.contains("unwrap_option_i32(some, 0LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_result_Option_i32_str");
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_option_i32");
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T unwrap_result"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("generic_ufc_function.zen"));
-    assert!(c_source.contains("int32_t id_i32(int32_t value)"));
-    assert!(c_source.contains("id_i32(12LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "id_i32");
-    assert!(!c_source.contains("id(12LL)"));
-    assert!(!c_source.contains("T id"));
 }
